+++ app/renderer/src/main/src/pages/invoker/LogCharts/LogCharts.module.scss
@@ -1,6 +1,3 @@
-%charts {
-    height: 360px;
-}
 %display-flex-center {
     display: flex;
     align-items: center;
@@ -12,30 +9,63 @@
 }
 .graph-xAxis-list {
     @extend %display-flex-center;
-    gap: 8px 20px;
+    gap: 12px 20px;
     border-bottom: 1px solid var(--yakit-border-color);
     padding-bottom: 8px;
     flex-wrap: wrap;
     .graph-xAxis-list-item {
         @extend %display-flex-center;
+        align-items: flex-start;
         gap: 8px;
         color: var(--yakit-body-text-color);
+        line-height: 16px;
     }
     .circle {
         width: 12px;
         height: 12px;
         border-radius: 50%;
+        flex-shrink: 0;
+        margin-top: 2px;
     }
 }
 
 .bar-graph-charts {
-    @extend %charts;
+    flex: 1;
+    height: 360px;
 }
-
-.pie-graph-charts {
-    @extend %charts;
-}
-
 .line-graph-charts {
-    @extend %charts;
+    flex: 1;
+    height: 360px;
+}
+.pie-graph-charts {
+    min-height: 420px;
+    padding: 24px;
+    min-width: 400px;
+}
+
+.pie-tooltip {
+    display: inline-flex;
+    flex-direction: column;
+    padding: 12px 16px;
+    max-width: 120px;
+}
+.pie-tooltip-name {
+    color: #ccd2de;
+    text-align: center;
+
+    font-size: 16px;
+    font-style: normal;
+    font-weight: 400;
+    line-height: 18px;
+}
+.pie-tooltip-value {
+    color: #fff;
+    text-align: center;
+
+    /* Headings/H5 */
+    font-family: Inter;
+    font-size: 18px;
+    font-style: normal;
+    font-weight: 700;
+    line-height: 24px; /* 133.333% */
 }
