+++ app/renderer/src/main/src/pages/invoker/LogCharts/LogCharts.tsx
@@ -5,7 +5,7 @@ import React, {useRef, useEffect, useState} from "react"
 import styles from "./LogCharts.module.scss"
 import {logChartsColorList} from "./constant"
 import {EChartsOption} from "@/pages/risks/VulnerabilityLevelPie/VulnerabilityLevelPieType"
-import {getBarOption, getLineOption} from "./chartsOption"
+import {getBarOption, getLineOption, getPieOption} from "./chartsOption"
 
 const logChartsColorListLength = logChartsColorList.length
 
@@ -13,9 +13,13 @@ interface LogChartsProps {
     type: "bar" | "line" | "pie"
     graphData: GraphData
 }
+interface LegendListProps {
+    value?: number
+    key: string
+}
 const LogCharts: React.FC<LogChartsProps> = React.memo((props) => {
     const {type, graphData} = props
-    const [legendList, setLegendList] = useState<string[]>([])
+    const [legendList, setLegendList] = useState<LegendListProps[]>([])
     const chartRef = useRef<HTMLDivElement>(null)
     const graphRef = useRef<echarts.ECharts>()
     const optionRef = useRef<EChartsOption>({})
@@ -23,7 +27,7 @@ const LogCharts: React.FC<LogChartsProps> = React.memo((props) => {
         if (!graphRef.current) {
             graphRef.current = echarts.init(chartRef.current)
         }
-        let newLegendList: string[] = []
+        let newLegendList: LegendListProps[] = []
         let option: EChartsOption = {}
         switch (type) {
             case "bar":
@@ -32,15 +36,30 @@ const LogCharts: React.FC<LogChartsProps> = React.memo((props) => {
             case "line":
                 option = getLineOption(graphData)
                 break
+            case "pie":
+                option = getPieOption(graphData)
+                break
             default:
                 break
         }
-        if (Array.isArray(option.series)) {
-            newLegendList = (option.series || []).map((ele) => ele.name as string)
+        switch (type) {
+            case "bar":
+            case "line":
+                if (Array.isArray(option.series)) {
+                    newLegendList = (option.series || []).map((ele) => ({key: ele.name as string}))
+                }
+                break
+            case "pie":
+                newLegendList = graphData.data
+                break
+            default:
+                break
         }
+
         optionRef.current = option
         graphRef.current.setOption(optionRef.current)
         setLegendList(newLegendList)
+
         return () => {
             if (graphRef.current) {
                 graphRef.current.dispose()
@@ -53,14 +72,14 @@ const LogCharts: React.FC<LogChartsProps> = React.memo((props) => {
             {legendList.length > 1 && (
                 <div className={styles["graph-xAxis-list"]}>
                     {legendList.map((item, index) => (
-                        <div key={item} className={styles["graph-xAxis-list-item"]}>
+                        <div key={item.key} className={styles["graph-xAxis-list-item"]}>
                             <div
                                 className={classNames(
                                     styles["circle"],
                                     `color-bg-${logChartsColorList[index % logChartsColorListLength]?.name || "purple"}`
                                 )}
                             />
-                            {item}
+                            <div>{item.key}</div>
                         </div>
                     ))}
                 </div>
@@ -68,8 +87,8 @@ const LogCharts: React.FC<LogChartsProps> = React.memo((props) => {
             <div
                 className={classNames({
                     [styles["bar-graph-charts"]]: type === "bar",
-                    [styles["pie-graph-charts"]]: type === "pie",
-                    [styles["line-graph-charts"]]: type === "line"
+                    [styles["line-graph-charts"]]: type === "line",
+                    [styles["pie-graph-charts"]]: type === "pie"
                 })}
                 ref={chartRef}
             ></div>
